Switch from Double to Decimal to handle optional values
Also corrected the logic when calculating and storing the historical 1RM data for each exercise

+++ OneRepMax/ExercisesList/ExercisesListViewModel.swift
@@ -11,7 +11,7 @@ import Foundation
 struct ExerciseListItem: Identifiable {
     let id: ObjectIdentifier
     let name: String
-    let oneRepMax: Double?
+    let oneRepMax: Decimal?
 }
 
 protocol ExercisesListViewModelProtocol: ObservableObject {
